requirements.txt
Js2Py==0.70
lxml==4.5.0
requests==2.23.0
click==7.1.2
pycryptodome==3.9.7
